pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "unified_assist"
version = "0.1.0"
description = "A Python super assistant runtime inspired by Claude Code"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["PyYAML>=6.0"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
